Revert "add modal for landing page support button (#10)"
This reverts commit 50137da6cc912a642812ac41615df800ab47ddf6.

revert contact modal
