+++ features/layout/landing-container/index.ts
@@ -1,4 +1,3 @@
 export { LandingHeader } from "./landing-header";
 export { LandingMenuItemLink } from "./landing-menu-item-link";
 export { LandingPageContainer } from "./landing-page-container";
-export { ContactButton } from "./contact-button";
